+++ app/composables/render/update/utils.ts
@@ -1,5 +1,6 @@
 import type { DataAirport } from '~/composables/render/storage';
 import { getActiveRunways } from '~/utils/shared/runway-detection';
+import type { VatsimController } from '~/types/data/vatsim';
 
 export function updateAirportAtisConfig(airport: DataAirport) {
     const atises: DataAirport['atis'] = {};
@@ -43,3 +44,5 @@ export function updateAirportAtisConfig(airport: DataAirport) {
         };
     }
 }
+
+export const debugControllers = useLocalStorage<VatsimController[]>('debug-controllers', []);